Schema introspection with auth headers blows up
Schema introspection does not have any fields with it so we should short circuit instead of blowing up on `field.type`. This allows the schema to be introspected when auth headers are set in clients like graphql playground or insomnia.

This is just a quick fix I found, not sure if it is 100% correct!